+++ services/resources/mgmt/2018-05-01/resources/models.go
@@ -2101,7 +2101,8 @@ type ProviderResourceType struct {
 	// Aliases - The aliases that are supported by this resource type.
 	Aliases *[]AliasType `json:"aliases,omitempty"`
 	// APIVersions - The API version.
-	APIVersions *[]string `json:"apiVersions,omitempty"`
+	APIVersions  *[]string      `json:"apiVersions,omitempty"`
+	ZoneMappings *[]ZoneMapping `json:"zoneMappings,omitempty"`
 	// Properties - The properties.
 	Properties map[string]*string `json:"properties"`
 }
@@ -2120,6 +2121,9 @@ func (prt ProviderResourceType) MarshalJSON() ([]byte, error) {
 	}
 	if prt.APIVersions != nil {
 		objectMap["apiVersions"] = prt.APIVersions
+	}
+	if prt.ZoneMappings != nil {
+		objectMap["zoneMappings"] = prt.ZoneMappings
 	}
 	if prt.Properties != nil {
 		objectMap["properties"] = prt.Properties
@@ -2549,4 +2553,11 @@ func (future *ValidateMoveResourcesFuture) result(client Client) (ar autorest.Re
 	}
 	ar.Response = future.Response()
 	return
+}
+
+// ZoneMapping ...
+type ZoneMapping struct {
+	// Location - The location of the zone mapping.
+	Location *string   `json:"location,omitempty"`
+	Zones    *[]string `json:"zones,omitempty"`
 }