+++ services/resources/mgmt/2019-10-01/resources/models.go
@@ -2822,7 +2822,8 @@ type ProviderResourceType struct {
 	// Aliases - The aliases that are supported by this resource type.
 	Aliases *[]Alias `json:"aliases,omitempty"`
 	// APIVersions - The API version.
-	APIVersions *[]string `json:"apiVersions,omitempty"`
+	APIVersions  *[]string      `json:"apiVersions,omitempty"`
+	ZoneMappings *[]ZoneMapping `json:"zoneMappings,omitempty"`
 	// Capabilities - The additional capabilities offered by this resource type.
 	Capabilities *string `json:"capabilities,omitempty"`
 	// Properties - The properties.
@@ -2843,6 +2844,9 @@ func (prt ProviderResourceType) MarshalJSON() ([]byte, error) {
 	}
 	if prt.APIVersions != nil {
 		objectMap["apiVersions"] = prt.APIVersions
+	}
+	if prt.ZoneMappings != nil {
+		objectMap["zoneMappings"] = prt.ZoneMappings
 	}
 	if prt.Capabilities != nil {
 		objectMap["capabilities"] = prt.Capabilities
@@ -3467,4 +3471,11 @@ type WhatIfPropertyChange struct {
 	After interface{} `json:"after,omitempty"`
 	// Children - Nested property changes.
 	Children *[]WhatIfPropertyChange `json:"children,omitempty"`
+}
+
+// ZoneMapping ...
+type ZoneMapping struct {
+	// Location - The location of the zone mapping.
+	Location *string   `json:"location,omitempty"`
+	Zones    *[]string `json:"zones,omitempty"`
 }